+++ packages/driver/cypress/e2e/issues/28187.cy.js
@@ -1,0 +1,6 @@
+it('hidden check', () => {
+  cy.visit('/fixtures/issue-28187.html')
+  cy.get('details')
+  .children('div')
+  .should('not.be.visible')
+})